+++ include/osurs/io.h
@@ -12,48 +12,64 @@
 #ifndef OSURS_IO_H_
 #define OSURS_IO_H_
 
+#include <limits.h>
 #include <osurs/network.h>
 #include <stdio.h>
+#include <unistd.h>
 
 /**
  * @brief Print node.
  *
- * @param Node The node struct to print.
+ * @param node The node struct to print.
  */
 void print_node(Node *node);
 
 /**
+ * @brief Print composition.
+ *
+ * @param composition The composition struct to print.
+ */
+void print_composition(Composition *composition);
+
+/**
+ * @brief Print vehicle.
+ *
+ * @param vehicle The vehicle struct to print.
+ */
+void print_vehicle(Vehicle *vehicle);
+
+/**
  * @brief Print stop.
  *
- * @param Stop The stop struct to print.
+ * @param stop The stop struct to print.
  */
 void print_stop(Stop *stop);
 
 /**
  * @brief Print trip.
  *
- * @param Trip The trip struct to print.
+ * @param trip The trip struct to print.
  */
 void print_trip(Trip *trip);
 
 /**
  * @brief Print route.
  *
- * @param Route The route struct to print.
+ * @param route The route struct to print.
  */
 void print_route(Route *route);
 
 /**
  * @brief Print network.
  *
- * @param Network The network struct to print.
+ * @param network The network struct to print.
  */
 void print_network(Network *network);
 
 /**
  * @brief Print connection.
  *
- * @param Connection The connection struct to print.
+ * @param connection The connection struct to print.
  */
 void print_connection(Connection *connection);
 
@@ -63,5 +79,34 @@ void print_connection(Connection *connection);
  * @param reservation The reservation struct to print.
  */
 void print_reservation(Reservation *reservation);
+
+/**
+ * @brief Prints the current working dirctory.
+ *
+ * @return 1 if success, 0 if failure.
+ */
+int print_cwd();
+
+/**
+ * @brief Function print file content to the console.
+ *
+ * @param filename The path to the file.
+ * @return 1 if success, 0 if failure.
+ */
+int print_file(const char *filename);
+
+/**
+ * @brief Import a MATSim transit schedule.
+ *
+ * Import a MATSim transit schedule from a transitSchedule.xml and  vehicles.xml
+ * file.
+ *
+ * @param network Pointer to an empty network.
+ * @param schedule_file File path to the transit schedule file.
+ * @param vehicle_file File path to the vehicles file.
+ * @return 1 if success, 0 if failure.
+ */
+int import_matsim(Network *network, const char *schedule_file,
+                  const char *vehicle_file);
 
 #endif  // OSURS_IO_H_