+++ include/osurs/network.h
@@ -21,8 +21,6 @@
 
 // Constructor-like methods
 
-// Create a Network struct
-
 /**
  * @brief Create a new nework
  *
@@ -54,29 +52,86 @@ Network *new_network();
 Node *new_node(Network *network, const char *id, double x, double y);
 
 /**
+ * @brief Create a new composition and add it to the network.
+ *
+ * @param network The network to add the composition.
+ * @param id The id of the composition.
+ * @param seats The seat capacity of the composition.
+ * @return A pointer to the newly allocated node (Composition*).
+ */
+Composition *new_composition(Network *network, const char *id, int seats);
+
+/**
+ * @brief Create a new vehicle and add it to the network.
+ *
+ * @param network The network to add the vehicle.
+ * @param id The id of the vehicle.
+ * @param composition The composition the vehicle has.
+ * @return A pointer to the newly allocated vehicle (Vehicle*).
+ */
+Vehicle *new_vehicle(Network *network, const char *id,
+                     Composition *composition);
+
+/**
  * @brief Create a new route and add it to the network.
  *
  * @param network The network to add the route.
+ * @param id The identifier of the route.
  * @param nodes A pointer to an array of nodes, which define the stops of the
  * route (n=route_size).
- * @param times A pointer to an array of travel times, which define the travel
- * time between the nodes (n=route_size-1).
+ * @param arrival_offsets A pointer to an array of arrival offset times from the
+ * root stop. The offsets define the travel times between the departure of the
+ * root stop and the arrival at the current stop (n=route_size).
+ * @param departure_offsets A pointer to an array of departure offset times from
+ * the root stop. The offsets define the travel times between the departure of
+ * the root stop and the departure at the current stop (n=route_size).
  * @param route_size The number of nodes (=stops) in the route.
+ * @param trip_ids The identifier of the trips on the route (n=trip_size).
  * @param departures The departure times at the first node (=root_stop) of the
  * route in seconds after midnight (00:00:00). Every departure time will define
- * a new trip, which starts traveling along the route (n = trip_size)
- * @param capacities This will be replaced with vehicles!
+ * a new trip, which starts traveling along the route (n=trip_size)
+ * @param vehicles An array of vehicles for each trip (n=trip_size).
  * @param trip_size The number of departures (=trips) on the route.
  * @return A pointer to the newly allocated node (Route*).
  */
-Route *new_route(Network *network, Node *nodes[], int times[],
-                 size_t route_size, int departures[], int capacities[],
-                 size_t trip_size);
+Route *new_route(Network *network, const char *id, Node *nodes[],
+                 int arrival_offsets[], int departure_offsets[],
+                 size_t route_size, const char *trip_ids[], int departures[],
+                 Vehicle *vehicles[], size_t trip_size);
+
+// Getters
+
+/**
+ * @brief Get the node struct.
+ *
+ * @param network An network to get the node from.
+ * @param id The identifier of the node.
+ * @return Returns the node.
+ */
+Node *get_node(Network *network, const char *id);
+
+/**
+ * @brief Get the vehicle struct.
+ *
+ * @param network An network to get the vehicle from.
+ * @param id The identifier of the vehicle.
+ * @return Returns the vehicle.
+ */
+Vehicle *get_vehicle(Network *network, const char *id);
+
+/**
+ * @brief Get the composition struct.
+ *
+ * @param network An network to get the composition from.
+ * @param id The identifier of the composition.
+ * @return Returns the composition.
+ */
+Composition *get_composition(Network *network, const char *id);
 
 // Destructor-like methods
 
 /**
- * @brief Delete a network
+ * @brief Delete a network.
  *
  * All objects of the network are located on the heap and are directly or
  * indirectly linked to the network structure. If the memory of the network is