+++ src/reserve.c
@@ -37,7 +37,7 @@ Connection *new_connection(const Node *orig, const Node *dest, int time) {
 
     // Mark start of the chain
     conn->next = NULL;
-    conn->last = NULL;
+    conn->prev = NULL;
 
     // Match and search on equal routes
     for (size_t i = 0; i < orig->route_counter; ++i) {
@@ -50,13 +50,13 @@ Connection *new_connection(const Node *orig, const Node *dest, int time) {
     }
 
     // Check for no results
-    if (conn->last == NULL && conn->next == NULL) {
+    if (conn->prev == NULL && conn->next == NULL) {
         free(conn);
         return NULL;
     }
 
     // Avoid memory leak and delete last allocated empty connection
-    conn->last->next = NULL;
+    conn->prev->next = NULL;
     free(conn);
 
     return root_conn;
@@ -74,7 +74,7 @@ int check(Connection *connection, int seats, int *trip_count) {
 
     // Check available seats over on all stops
     Stop *curr_stop = connection->orig;
-    int capacity = connection->trip->capacity;
+    int capacity = connection->trip->vehicle->composition->seats;
     while (1) {
         if (seats > (capacity - curr_stop->reserved[*trip_count])) return 0;
         // Stop if destination is reached
@@ -107,7 +107,7 @@ int new_reservation(Connection *connection, int seats) {
 void delete_connection(Connection *connection) {
     if (connection == NULL) return;
 
-    Connection *last_conn;
+    Connection *prev_conn;
     Connection *next_conn;
     Connection *curr_conn;
     Connection *root_conn = connection;
@@ -120,13 +120,13 @@ void delete_connection(Connection *connection) {
         free(curr_conn);
         curr_conn = next_conn;
     }
-    // Check for last
-    curr_conn = root_conn->last;
+    // Check for prev
+    curr_conn = root_conn->prev;
     while (1) {
         if (curr_conn == NULL) break;
-        last_conn = curr_conn->last;
+        prev_conn = curr_conn->prev;
         free(curr_conn);
-        curr_conn = last_conn;
+        curr_conn = prev_conn;
     }
     // Finally free root
     free(root_conn);
@@ -140,8 +140,10 @@ int min(int a, int b) { return (a > b) ? b : a; }
 // Iterate to stop of origin node on trip
 Stop *iterate_to_orig(Stop *curr_stop, const Node *orig, int *departure) {
     while (1) {
-        if (curr_stop->node == orig) return curr_stop;
-        *departure += curr_stop->time_to_next;
+        if (curr_stop->node == orig) {
+            *departure += curr_stop->departure_offset;
+            return curr_stop;
+        }
         curr_stop = curr_stop->next;
     }
 }
@@ -152,8 +154,10 @@ Stop *iterate_to_dest(Stop *curr_stop, const Node *dest, int *arrival,
     while (1) {
         *available =
             min(*available, capacity - curr_stop->reserved[trip_count]);
-        if (curr_stop->node == dest) return curr_stop;
-        *arrival += curr_stop->time_to_next;
+        if (curr_stop->node == dest) {
+            *arrival += curr_stop->arrival_offset;
+            return curr_stop;
+        }
         curr_stop = curr_stop->next;
     }
 }
@@ -178,7 +182,7 @@ Connection *search_trip(Connection *conn, const Node *orig, const Node *dest,
     // Set departure at origin as base for calculation of the arrival at origin.
     arrival = departure;
     dest_stop = iterate_to_dest(orig_stop, dest, &arrival, &available,
-                                trip->capacity, trip_count);
+                                trip->vehicle->composition->seats, trip_count);
 
     // Set in values of found connection
     *conn_count += 1;
@@ -190,11 +194,11 @@ Connection *search_trip(Connection *conn, const Node *orig, const Node *dest,
     conn->available = available;
 
     // Allocate next connection on heap; Set pointer to current connection as
-    // last connection of next connection.
+    // prev connection of next connection.
     conn->next = (Connection *)malloc(sizeof(Connection));
-    Connection *conn_last = conn;
+    Connection *conn_prev = conn;
     conn = conn->next;
-    conn->last = conn_last;
+    conn->prev = conn_prev;
 
     return conn;
 }
@@ -207,8 +211,11 @@ Connection *search_route(Connection *conn, const Node *orig, const Node *dest,
     int conn_count = 0;
 
     while (1) {
-        conn = search_trip(conn, orig, dest, time, curr_trip, route->root_stop,
-                           trip_count, &conn_count);
+        // Search trip if not already arrived at terminal
+        if (curr_trip->arrival > time) {
+            conn = search_trip(conn, orig, dest, time, curr_trip,
+                               route->root_stop, trip_count, &conn_count);
+        }
         if (conn_count >= cutoff || curr_trip->next == NULL) return conn;
         curr_trip = curr_trip->next;
         ++trip_count;